backend_python/requirements.txt
fastapi
uvicorn[standard]
python-socketio
sqlalchemy
pymysql
python-dotenv
aiomysql
cryptography
